adversarial fiddling

+++ Code/Models/Attention_seq2seq.py
@@ -388,7 +388,8 @@ class _Seq2Seq(nn.Module):
         if adversarial == True:
             hidden = hidden_.clone() + torch.from_numpy(
                 np.random.normal(0, 0.01, size = hidden_.shape)
-                ).to(self.device).double()
+                ).to(self.device).float()
+            hidden = hidden.float()
         else:
             hidden = hidden_
